Ian (#17)
* Seed Planting and Prototype Harvesting

Modified GameScreen to add a prototype shop pumpkin button. When this button is clicked, it sends the seed of the button down a chain from gamescreen -> plot -> tile to prepare tiles to change their seed. Now when the text is clicked the UITtile plants the selected seed.

* Plant & global variables

Added globalState class that includes global variables, modified the plant methods to rely on the global variables