Ian (#26)
* Added final timer

-Updated GameScreen to forcibly end the game after 10 minutes

* Update tile images

Xander added functionality to update the current image of the tile based on the growth stage.

* Add harvest button

Added a harvest button to gameScreen that reads "stop planting" that appears when player is planting and disappears if the player is in harvest mode.

* Harvesting adds money

Harvesting a seed adds the worth of the seed to the PlayerData.money through a function in PlayerData called addMoney()

* Money Label

Added a money label on GameScreen that shows the current amount of money the player has

* bug fixes

fixed a bug with harvestSeed where it would start at growth stage 2

* Progress bar

Replaced click button underneath tile with a progress bar the indiciates how far along tile is to next growth stage. If the tile is rotten, the progress bar stays full. Progress bar resets when harvested.